requirements.txt
altair==5.2.0
numpy==1.24.4
pandas==2.0.3
PyYAML==6.0.1
streamlit==1.31.1
streamlit_authenticator==0.3.1
openpyxl